Staging (#28)
* feat(auth): add id property to User model

* feat(auth): add id property to User model

* refactor: Update start:prod script in package.json

The start:prod script in package.json was updated to use the correct file path for the main script.

Deleted migration file for updating User model

The migration.sql file for updating the User model was deleted.

Create migration file for initializing User model

A new migration.sql file was created to initialize the User model with the required columns and constraints.

* feat(auth): Add redirect to frontend with token after successful Google authentication

* feat(auth): Implement JWT token generation in GoogleStrategy

* feat(auth): Add redirect to frontend with JWT token in URL

* refactor(main): Remove unused tag in Swagger configuration

* Feature/oauth google redirection (#23)

* * merge(auth): add id property to User model

* feat(auth): add id property to User model

* feat(auth): add id property to User model

* refactor: Update start:prod script in package.json

The start:prod script in package.json was updated to use the correct file path for the main script.

Deleted migration file for updating User model

The migration.sql file for updating the User model was deleted.

Create migration file for initializing User model

A new migration.sql file was created to initialize the User model with the required columns and constraints.

* feat(auth): Add redirect to frontend with token after successful Google authentication

* feat(auth): Implement JWT token generation in GoogleStrategy

* feat(auth): Add redirect to frontend with JWT token in URL

* refactor(main): Remove unused tag in Swagger configuration

* refactor(main): Update @types/jest to version 29.5.13

* Feature/jwt (#25)

* feat(auth): add JwtAuthGuard and JwtStrategy

* fix(auth): Update JwtAuthGuard and JwtStrategy imports

* refactor(auth): delete all comments

* feat(auth): add class-validator package

* feat(auth): add SignUpDto and signUp endpoint

* fix(auth): Swagger translate to english

* refactor(auth): Update login endpoint path to 'log-in'

* feat(auth): Add GoogleStrategy and JWT token generation

* fix merge

* tests(google): add tests for google oauth

* tests(auth): add tests for
- validateUser
- login

* tests(auth) add signup test

* tests(auth): add tests for all get

* tests(auth): add tests for :
- update user
- delete user
- validate oauth login

* tests(auth): add tests for AuthController

* test(auth): add googleAuthRedirect test

Added a new test case for the `googleAuthRedirect` method in the `AuthController` class. The test ensures that the method redirects to Google OAuth with the provided token. This test enhances the test coverage for the authentication functionality.

* tests(auth): add tests for
- module
- hashPassword
- jwt

* tests(e2e): add tests for e2e usage